[METAED-1486] warnings should be on extensions only
