+++ packages/metaed-plugin-edfi-odsapi/src/validator/ValidatorList.ts
@@ -4,13 +4,13 @@ import { validate as includingRequiredPropertiesInExtendedAssociationsProhibited
 import { validate as includingRequiredPropertiesInExtendedDomainEntitiesProhibited } from './UpcomingImprovements/IncludingRequiredPropertiesInExtendedDomainEntitiesProhibited';
 import { validate as subclassingAnyAssociationExceptGeneralStudentProgramAssociationIsUnsupported } from './UpcomingImprovements/SubclassingAnyAssociationExceptGeneralStudentProgramAssociationIsUnsupported';
 import { validate as subclassingAnyDomainEntityExceptEducationOrganizationIsUnsupported } from './UpcomingImprovements/SubclassingAnyDomainEntityExceptEducationOrganizationIsUnsupported';
-import { validate as ignoreRequiredChoiceProperty } from './OdsApiIgnoreRequiredChoiceProperty';
+import { validate as odsApiIgnoresRequiredOnChoicePropertiesInExtensions } from './OdsApiIgnoresRequiredOnChoicePropertiesInExtensions';
 
 export function validatorList(): Validator[] {
   return [
     includingRequiredPropertiesInExtendedAssociationsProhibited,
     includingRequiredPropertiesInExtendedDomainEntitiesProhibited,
-    ignoreRequiredChoiceProperty,
+    odsApiIgnoresRequiredOnChoicePropertiesInExtensions,
     subclassingAnyAssociationExceptGeneralStudentProgramAssociationIsUnsupported,
     subclassingAnyDomainEntityExceptEducationOrganizationIsUnsupported,
   ];
